Revert "chore: dynamic imports"
This reverts commit 476d8b4c3ad6afd608c3c54a9dcee6c6c3805c8a.
