+++ explorer/next.config.js
@@ -4,23 +4,6 @@ const nextConfig = {
   // Enable experimental features for better optimization
   experimental: {
     optimizeCss: true,
-    // Enable more aggressive optimizations
-    optimizePackageImports: [
-      '@nivo/core',
-      '@nivo/line',
-      '@nivo/bar',
-      '@nivo/pie',
-      'react-icons',
-      'ethers',
-      '@polkadot/extension-dapp',
-      '@polkadot/react-identicon',
-      'swiper',
-    ],
-    // Enable lazy compilation for better performance
-    lazyCompilation: {
-      entries: false,
-      imports: true,
-    },
   },
   // Enable compression
   compress: true,
@@ -43,7 +26,7 @@ const nextConfig = {
     ],
   },
   // Webpack optimizations
-  webpack: (config, { isServer, dev }) => {
+  webpack: (config, { isServer }) => {
     // SVG loader
     config.module.rules.push({
       test: /\.svg$/,
@@ -59,85 +42,6 @@ const nextConfig = {
 
     if (!isServer) {
       config.resolve.fallback.fs = false
-
-      // Enable tree shaking and module concatenation
-      config.optimization = {
-        ...config.optimization,
-        usedExports: true,
-        sideEffects: false,
-        concatenateModules: true,
-        splitChunks: {
-          chunks: 'all',
-          maxInitialRequests: 25,
-          minSize: 20000,
-          cacheGroups: {
-            // Separate vendor chunks for better caching
-            vendor: {
-              test: /[\\/]node_modules[\\/]/,
-              name: 'vendors',
-              chunks: 'all',
-              priority: 10,
-            },
-            // Separate large libraries
-            nivo: {
-              test: /[\\/]node_modules[\\/]@nivo[\\/]/,
-              name: 'nivo',
-              chunks: 'all',
-              priority: 20,
-              enforce: true,
-            },
-            polkadot: {
-              test: /[\\/]node_modules[\\/]@polkadot[\\/]/,
-              name: 'polkadot',
-              chunks: 'all',
-              priority: 20,
-              enforce: true,
-            },
-            ethers: {
-              test: /[\\/]node_modules[\\/]ethers[\\/]/,
-              name: 'ethers',
-              chunks: 'all',
-              priority: 20,
-              enforce: true,
-            },
-            swiper: {
-              test: /[\\/]node_modules[\\/]swiper[\\/]/,
-              name: 'swiper',
-              chunks: 'all',
-              priority: 20,
-              enforce: true,
-            },
-            reactIcons: {
-              test: /[\\/]node_modules[\\/]react-icons[\\/]/,
-              name: 'react-icons',
-              chunks: 'all',
-              priority: 20,
-              enforce: true,
-            },
-            // Separate form libraries
-            forms: {
-              test: /[\\/]node_modules[\\/](formik|yup)[\\/]/,
-              name: 'forms',
-              chunks: 'all',
-              priority: 15,
-            },
-            // Separate chart libraries
-            charts: {
-              test: /[\\/]node_modules[\\/](d3|@nivo)[\\/]/,
-              name: 'charts',
-              chunks: 'all',
-              priority: 15,
-            },
-            // Separate utility libraries
-            utils: {
-              test: /[\\/]node_modules[\\/](lodash|dayjs|clsx)[\\/]/,
-              name: 'utils',
-              chunks: 'all',
-              priority: 15,
-            },
-          },
-        },
-      }
     }
 
     return config
